Rename /lib/atomic to lib/swapper
